ok I'll explain
the major change that's pretty obvious is the time modifier on the sponsor page. Another change that you won't notice immediately is that I added a function called getData that gets the data from the google spreadsheet that holds the email for every entrant and what raffle they entered in. I will probably link the google form in the readme of this github but I used google forms as a substitute for any major backend. I also made it so when the user enters a raffle it updates on the google form and spreadsheet